requirements.txt
discord.py[voice]>=2.4.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
alembic>=1.13.0
google-generativeai>=0.3.0 